+++ algorithm_tester/__main__.py
@@ -1,3 +1,3 @@
-from algorithm_tester.solver_runner import main
+from algorithm_tester.tester_cli_interface import main
 
 main("tester")
